+++ lib/dlex.ex
@@ -10,6 +10,7 @@ defmodule Dlex do
 
   @type conn :: DBConnection.conn()
   @type uid :: String.t()
+  @type queries :: %{optional(:query | :condition) => String.t()}
 
   @timeout 15_000
   @default_keepalive :infinity
@@ -168,11 +169,13 @@ defmodule Dlex do
     * `:timeout` - Call timeout (default: `#{@timeout}`)
   """
 
-  @spec mutate(conn, iodata, iodata | map, Keyword.t()) ::
+  @spec mutate(conn, queries, iodata | map, Keyword.t()) ::
           {:ok, map} | {:error, Dlex.Error.t() | term}
 
-  def mutate(conn, query, statement, opts) do
-    query = %Query{type: Type.Mutation, statement: statement, condition: query}
+  def mutate(conn, queries, statement, opts) do
+    condition = Map.get(queries, :condition)
+    query = Map.get(queries, :query, "")
+    query = %Query{type: Type.Mutation, condition: condition, statement: statement, query: query}
 
     with {:ok, _, result} <- DBConnection.prepare_execute(conn, query, %{}, opts),
          do: {:ok, result}
@@ -181,24 +184,21 @@ defmodule Dlex do
   @doc """
   The same as `Dlex.mutate(conn, "", mutation, [])`
   """
-  def mutate(conn, statement), do: mutate(conn, "", statement, [])
+  def mutate(conn, statement), do: mutate(conn, %{}, statement, [])
 
   @doc """
-  The same as `Dlex.mutate(conn, query, mutation, [])` or `Dlex.mutate(conn, "", mutation, opts)`
+  The same as `Dlex.mutate(conn, queries, mutation, [])`.
   """
-  def mutate(conn, query, statement) when not is_list(statement),
-    do: mutate(conn, query, statement, [])
-
   def mutate(conn, statement, opts) when is_list(opts),
-    do: mutate(conn, "", statement, opts)
+    do: mutate(conn, %{}, statement, opts)
 
   @doc """
   Runs a mutation and returns the result or raises `Dlex.Error` if there was an error.
-  See `mutate/4`.
+  See `mutate/5`.
   """
-  @spec mutate!(conn, iodata, iodata | map, Keyword.t()) :: map | no_return
-  def mutate!(conn, query, statement, opts) do
-    case mutate(conn, query, statement, opts) do
+  @spec mutate!(conn, queries, iodata, Keyword.t()) :: map | no_return
+  def mutate!(conn, queries, statement, opts) do
+    case mutate(conn, queries, statement, opts) do
       {:ok, result} -> result
       {:error, err} -> raise err
     end
@@ -206,10 +206,11 @@ defmodule Dlex do
 
   @doc """
   Runs a mutation and returns the result or raises `Dlex.Error` if there was an error.
-  See `mutate/3`.
+  See `mutate/4`.
   """
-  def mutate!(conn, query_or_statement, statement_or_opts) do
-    case mutate(conn, query_or_statement, statement_or_opts) do
+  @spec mutate!(conn, iodata | map, Keyword.t()) :: map | no_return
+  def mutate!(conn, statement, opts) do
+    case mutate(conn, statement, opts) do
       {:ok, result} -> result
       {:error, err} -> raise err
     end
@@ -220,7 +221,7 @@ defmodule Dlex do
   See `mutate/2`.
   """
   def mutate!(conn, statement) do
-    case mutate(conn, nil, statement, []) do
+    case mutate(conn, %{}, statement, []) do
       {:ok, result} -> result
       {:error, err} -> raise err
     end
@@ -264,12 +265,12 @@ defmodule Dlex do
 
   """
   @spec delete(conn, iodata | map, Keyword.t()) :: {:ok, map} | {:error, Dlex.Error.t() | term}
-  def delete(conn, condition, statement, opts) do
+  def delete(conn, query, statement, opts) do
     query = %Query{
       type: Type.Mutation,
       sub_type: :deletion,
       statement: statement,
-      condition: condition
+      query: query
     }
 
     with {:ok, _, result} <- DBConnection.prepare_execute(conn, query, %{}, opts),