+++ test/dlex/node_test.exs
@@ -3,22 +3,17 @@ defmodule Dlex.NodeTest do
 
   alias Dlex.User
 
-  setup_all do
-    {:ok, pid} = Dlex.start_link(pool_size: 4)
-    %{pid: pid}
-  end
-
   describe "schema generation" do
     test "basic" do
-      assert "type.user" == User.__schema__(:source)
+      assert "user" == User.__schema__(:source)
       assert :string == User.__schema__(:type, :name)
       assert :integer == User.__schema__(:type, :age)
       assert [:name, :age, :friends] == User.__schema__(:fields)
     end
 
     test "alter" do
-      assert %{ 
-              "schema" => [
+      assert %{
+               "schema" => [
                  %{
                    "index" => true,
                    "predicate" => "user.name",
@@ -27,18 +22,18 @@ defmodule Dlex.NodeTest do
                  },
                  %{"predicate" => "user.age", "type" => "int"},
                  %{"predicate" => "user.friends", "type" => "uid"}
-              ],
-              "types" => [
-                %{
-                  "fields" => [
-                    %{"name" => "user.friends", "type" => "uid"},
-                    %{"name" => "user.age", "type" => "integer"},
-                    %{"name" => "user.name", "type" => "string"}
-                  ],
-                  "name" => "type.user"
-                }
-              ]
-            } == User.__schema__(:alter)
+               ],
+               "types" => [
+                 %{
+                   "fields" => [
+                     %{"name" => "user.friends", "type" => "uid"},
+                     %{"name" => "user.age", "type" => "integer"},
+                     %{"name" => "user.name", "type" => "string"}
+                   ],
+                   "name" => "user"
+                 }
+               ]
+             } == User.__schema__(:alter)
     end
 
     test "transformation callbacks" do